refactor: replace old MainPage with new MainPage2 implementation
- Backup original MainPage to MainPageOld.jsx\n- Promote MainPage2.jsx content to MainPage.jsx\n- Delete MainPage2.jsx\n- Update App.js routes
