+++ closzIT-front/src/App.js
@@ -5,8 +5,8 @@ import AuthCallbackPage from './pages/Login/AuthCallbackPage';
 import UserProfileSetup1 from './pages/UserProfileSetup/UserProfileSetup1';
 import UserProfileSetup2 from './pages/UserProfileSetup/UserProfileSetup2';
 import UserProfileSetup3 from './pages/UserProfileSetup/UserProfileSetup3';
+import MainPageOld from './pages/Main/MainPageOld';
 import MainPage from './pages/Main/MainPage';
-import MainPage2 from './pages/Main/MainPage2';
 import MyPage from './pages/MyPage/MyPage';
 import RegisterPage from './pages/Register/RegisterPage';
 import LabelingPage from './pages/Labeling/LabelingPage';
@@ -34,8 +34,8 @@ function App() {
           <Route path="/setup2" element={<UserProfileSetup2 />} />
           <Route path="/setup3" element={<UserProfileSetup3 />} />
           <Route path="/setup/profile3" element={<UserProfileSetup3 />} />
-          <Route path="/main" element={<MainPage2 />} />
-          <Route path="/main_ex" element={<MainPage />} />
+          <Route path="/main" element={<MainPage />} />
+          <Route path="/main_old" element={<MainPageOld />} />
           {/* <Route path="/main2" element={<MainPage2 />} /> */}
           <Route path="/mypage" element={<MyPage />} />
           <Route path="/register" element={<RegisterPage />} />
